requirements.txt
beautifulsoup4==4.10.0
crochet
Pillow
celery==5.2.3
Django==3.2.11
django-admin-thumbnails
numpy==1.21.5
requests==2.26.0
Scrapy==2.5.1
django-filter
django-debug-toolbar
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
django-background-tasks==1.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-compat==1.0.15
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
django-templated-mail==1.1.1
gunicorn==20.1.0
jquery.javascript==1.0.0
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
Scrapy==2.5.1
unicorn==1.0.3
whitenoise
dj-database-url
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
transaction==3.0.1
Twisted==21.7.0
typing_extensions==4.0.1
tzdata==2021.5
unicorn==1.0.3
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
yarg==0.1.9
zope.browser==2.3
zope.browserpage==4.4.0
zope.component==5.0.1
zope.configuration==4.4.0
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.exceptions==4.4
zope.hookable==5.1.0
zope.i18n==4.9.0
zope.i18nmessageid==5.0.1
zope.interface==5.4.0
zope.location==4.2
zope.pagetemplate==4.6.0
zope.proxy==4.5.0
zope.publisher==6.0.2
zope.schema==6.2.0
zope.security==5.1.1
zope.tal==4.5
zope.tales==5.1
zope.traversing==4.4.1
zope.viewlet==4.2.1

